+++ packages/storage-plus/src/lib.rs
@@ -1,10 +1,8 @@
+mod helpers;
 mod keys;
-mod length_prefixed;
 mod map;
-mod namespace_helpers;
 mod path;
 mod prefix;
-mod type_helpers;
 
 pub use keys::{Prefixer, PrimaryKey};
 pub use map::Map;
