+++ packages/storage-plus/src/prefix.rs
@@ -3,9 +3,7 @@ use serde::de::DeserializeOwned;
 use serde::Serialize;
 use std::marker::PhantomData;
 
-use crate::length_prefixed::nested_namespaces_with_key;
-use crate::namespace_helpers::range_with_prefix;
-use crate::type_helpers::deserialize_kv;
+use crate::helpers::{deserialize_kv, nested_namespaces_with_key, range_with_prefix};
 use cosmwasm_std::{Order, StdResult, Storage, KV};
 
 pub struct Prefix<T>
